backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
google-genai>=1.0.0
Pillow>=11.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
opencv-python-headless>=4.8.0
